Build OCKC and execute OpenJCEPlus tests
The github action associated with this repository will now clone code
from the OpenCryptographyKitC open source project and build this code
as part of the github action.

Given that the OpenCryptographyKitC only builds the libraries associated
with the OpenJCEPlus provider we will only run those tests for the time
being.