Modify SymmetricCipher overlaps buffer checks
The `doFinal` and `update` methods in the SymmetricCipher class have
logic to detect if a input buffer and output buffer are overlapping
in any sort of way. If they are then a copy of the input is made to a
separate location to ensure a safe operation can then take place on the
data.

There is a logic problem where the calculation of overlap was incorrect
and used a `=` instead of a `+` to calculate if the input and output
overlapped. The comment above this code block was correct and remains
unmodified.

Signed-off-by: Jason Katonica <katonica@us.ibm.com>